requirements.txt
numpy
matplotlib
pillow
scikit-learn
customtkinter
packaging